pyproject.toml
[tool.poetry]
name = "prep"
version = "0.6.0"
description = ""
authors = ["fwf-upetrading <frederick.fillingham@upetrading.com>"]
readme = "README.md"

[tool.poetry.dependencies]
pandas = {extras = ["performance"], version = "^2.1.1"}
python = ">=3.10,<3.13"
sqlalchemy = "^2.0.19"
psycopg = {extras = ["binary"], version = "^3.1.9"}
sqlalchemy-json = "^0.6.0"
redis = {extras = ["hiredis"], version = "^4.6.0"}
azure-functions = "^1.15.0"
python-dateutil = "^2.8.2"
upedata = {path = "dependencies/upedata-0.6.2-py3-none-any.whl"}
requests = "^2.31.0"
ujson = "^5.8.0"
paramiko = "^3.3.1"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
pytest = "^7.4.0"
pipdeptree = "^2.13.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.11.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
